Added the conditions to show the current character if it equals the character in string at corresponding position
I created two variables. The first gives the input position. The second get's the character at input position.

Using those variables I created an if statement where, if the current character in input = the character in the string from word[] then do something.

For now I'm just logging the character in the console.