Changed code to add letters to document instead of words
I removed the ul since I'm not going to remove words from the document.

To show progression the letter will change in opacity or some other style to signify the user pressed the right key.

To add letters one by one, I used a for loop that goes through every character of an item and adds it to the document.

I left the progressionRender function at the bottom as this is what I'll be working on next.