Changed the style of spans for progression successfully but found an error
The colors change to green for right and red for wrong and the code works fine.

A problem I've set aside that has become apparent since I'm checking values letter by letter instead of by words is if the user types too fast, the progression function doesn't have time to run.

This may be a performance issue or an issue with the input element. I'll try switching to getting the key pressed instead of the input value.